+++ src/EFCore.Ydb/src/Metadata/Conventions/YdbConventionSetBuilder.cs
@@ -13,7 +13,6 @@ public class YdbConventionSetBuilder(
     {
         var coreConventions = base.CreateConventionSet();
         coreConventions.Add(new YdbStringAttributeConvention(Dependencies));
-        coreConventions.Add(new YdbValueGenerationConvention(Dependencies, RelationalDependencies));
         return coreConventions;
     }
 }